Convert more Redux code into classes
