+++ src/renderer/rpcClient/checkIfRpcIssueHasBeenSolved.ts
@@ -1,5 +1,5 @@
 import { Actions } from "_r/redux/Actions";
-import { store } from "_r/redux/reducers/store";
+import { store } from "_r/redux/StoreCreator";
 
 /**
  * This function resolves only when Bitcoin Core gives a non-error response.
